Cambios y agregacion de estados de la vista de los usuarios (#17)
* Feat modulo fabricantes (#12)

* FEAT: Implementacion modulo fabricantes

* FEAT: Modulo Fabricantes (#14)

* BUGFIX: Se soluciona el error de renderizado de las alertas

* FEAT: Implementacion modulo fabricantes

* BUGFIX: Se arreglan las dimensiones de la pagina del login en pantallas grandes

* BUGFIX: Actualizacion de mensaje de error al crear el fabricante

* feat:user id and fix:tabla clientes and css

* Feat modulo fabricantes (#19)

* FEAT: Implementacion modulo fabricantes

* BUGFIX: Se arreglan las dimensiones de la pagina del login en pantallas grandes

* BUGFIX: Actualizacion de mensaje de error al crear el fabricante

* FEAT: Vista individual de fabricantes y edicion de la informacion de fabricante


* REAFACTOR: Se refactorizan estilos y nombre de los archivos

---------
