+++ src/features/usuarios/components/EditUserForm/EditUserForm.module.css
@@ -1,0 +1,73 @@
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.modal {
+  background: #fff;
+  padding: 2rem;
+  border-radius: 12px;
+  box-shadow: 0 8px 20px rgba(0, 0, 0, 0.2);
+  min-width: fit-content;
+  width: 900px;
+}
+
+.modal h2{
+  color: var(--color-primario);
+  font-size: 2rem;
+  font-weight: 700;
+  text-align: center;
+  margin-bottom: 30px;
+}
+
+.form{
+  display: flex;
+  flex-direction: column;
+  gap: 50px;
+}
+
+.inputsContainer{
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  gap: 40px;
+}
+
+.inputsContainer .rightColumn, .leftColumn{
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.buttonSelector{
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
+
+.buttonSelector label{
+  font-weight: 700;
+  font-size: 1rem;
+  color: var(--color-primario);
+}
+
+.buttonSelector .buttonsContainer{
+  display: flex;
+  gap: 30px;
+  align-items: flex-start;
+}
+
+.buttons {
+  display: flex;
+  justify-content: center;
+  gap: 20px;
+}
+
