+++ src/features/usuarios/components/UserCard/UserCard.module.css
@@ -1,0 +1,34 @@
+/* Contenedor */
+
+.mainContainer{
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  width: 100%;
+}
+
+.box {
+  background: white;
+  padding: 30px;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+  min-width: fit-content;
+  width: 25dvw;
+}
+
+.infoContainer{
+  display: flex;
+  flex-direction: column;
+  gap: 5px;
+}
+
+.buttonGroup{
+  display: flex;
+  gap: 20px;
+}
+/* texto para que se vea un poco mejor */
+.texts {
+  color: #333;
+  font-weight: 600;
+  font-size: 1.1rem;
+}
