+++ src/pages/dashboard/index.astro
@@ -7,34 +7,3 @@ import "./index.css"
 </SidebarLayout>
 
 
-    <main>
-        <div class="contenedorTablaProductos">
-            <TablaProductos/>
-        </div>
-        <div class="peticionesYbarraproductosAnalisados">
-            <div class="peticiones">
-                <article class="peticionesPendientes">
-                    <h3>Peticiones de cambio pendientes</h3>
-                    <p>18</p>
-                </article>
-                <article class="peticionesListas">
-                    <h3>Peticiones de cambio listas</h3>
-                    <p>35</p>
-                </article>
-            </div>
-            <div class="barrraProductosAnalisados">
-                <GraficaBarras/>
-            </div>
-        </div>
-        <div class="tablaClientesyCerificados">
-            <div class="tblClientes">
-                <TablaClientes/>
-            </div>
-            <div class="graficaCircular">
-                <GraficaCircular/>
-            </div>
-            
-        </div>
-    </main>
-</SidebarLayout>
-
