+++ src/shared/components/buttons/select-button/SelectButton.module.css
@@ -1,80 +1,109 @@
 /* ---------------------------------------- */
 /* Estilos base del boton */
 /* ---------------------------------------- */
-.selectButton{
-    padding: 10px;
-    border: 3px solid;
-    border-radius: 10px;
-    background: none;
-    font-size: 16px;
-    font-weight: 700;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    gap: 10px;
-    transition: .25s all;
-    width: 100%;
+.selectButton {
+  padding: 10px;
+  border: 3px solid;
+  border-radius: 10px;
+  background: none;
+  font-size: 16px;
+  font-weight: 700;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+  transition: 0.25s all;
+  width: 100%;
 }
 
 /* Estilos para ionos dentro del boton */
-.selectButton svg{
-    height: 1.5em;
-    width: auto;
-    stroke-width: 1.5;
+.selectButton svg {
+  height: 1.5em;
+  width: auto;
+  stroke-width: 1.5;
 }
 
 /* ---------------------------------------- */
 /* Estilos variante azul claro */
 /* ---------------------------------------- */
-.selectButton.lightBlue{
-    color: #81BFFF;
+.selectButton.lightBlue {
+  color: #81bfff;
 }
 
-.selectButton.lightBlue:hover, .selectButton.lightBlue.selected{
-    color: white;
-    border-color: transparent;
-    background:  #81BFFF;
+.selectButton.lightBlue:hover,
+.selectButton.lightBlue.selected {
+  color: white;
+  border-color: transparent;
+  background: #81bfff;
 }
-
 
 /* ---------------------------------------- */
 /* Estilos variante azul oscuro */
 /* ---------------------------------------- */
-.selectButton.darkBlue{
-    color: var(--color-primario);
+.selectButton.darkBlue {
+  color: var(--color-primario);
 }
 
-.selectButton.darkBlue:hover, .selectButton.darkBlue.selected{
-    color: white;
-    border-color: transparent;
-    background:  var(--color-primario);
+.selectButton.darkBlue:hover,
+.selectButton.darkBlue.selected {
+  color: white;
+  border-color: transparent;
+  background: var(--color-primario);
 }
 
 /* ---------------------------------------- */
 /* Estilos variante naranja */
 /* ---------------------------------------- */
 
-.selectButton.orange{
-    color: #FFB146 ;
+.selectButton.orange {
+  color: #ffb146;
 }
 
-.selectButton.orange:hover, .selectButton.orange.selected{
-    color: white;
-    background: #FFB146;
-    border-color: transparent;
+.selectButton.orange:hover,
+.selectButton.orange.selected {
+  color: white;
+  background: #ffb146;
+  border-color: transparent;
 }
 
 /* ---------------------------------------- */
 /* Estilos variante gris */
 /* ---------------------------------------- */
 
-.selectButton.gray{
-    color: #5B6670 ;
+.selectButton.gray {
+  color: #5b6670;
 }
 
-.selectButton.gray:hover, .selectButton.gray.selected{
-    color: white;
-    background: #5B6670;
-    border-color: transparent;
+.selectButton.gray:hover,
+.selectButton.gray.selected {
+  color: white;
+  background: #5b6670;
+  border-color: transparent;
+}
+/* ---------------------------------------- */
+/* Estilos variante verde (Activo) */
+/* ---------------------------------------- */
+.selectButton.green {
+  color: var(--color-verde);
 }
 
+.selectButton.green:hover,
+.selectButton.green.selected {
+  color: white;
+  background: var(--color-verde);
+  border-color: transparent;
+}
+
+/* ---------------------------------------- */
+/* Estilos variante rojo (Inactivo) */
+/* ---------------------------------------- */
+.selectButton.red {
+  color: var(--color-rojo);
+}
+
+.selectButton.red:hover,
+.selectButton.red.selected {
+  color: white;
+  background: var(--color-rojo);
+  border-color: transparent;
+}
